+++ lib/Fhaculty/Graph/Algorithm/ShortestPath/Base.php
@@ -4,13 +4,52 @@ namespace Fhaculty\Graph\Algorithm\ShortestPath;
 
 use Fhaculty\Graph\Algorithm\BaseVertex;
 use Fhaculty\Graph\Walk;
-use Fhaculty\Graph\Exception\UnderflowException;
+use Fhaculty\Graph\Exception\OutOfBoundsException;
 use Fhaculty\Graph\Exception\InvalidArgumentException;
 use Fhaculty\Graph\Vertex;
 use Fhaculty\Graph\Edge\Base as Edge;
 use Fhaculty\Graph\Set\Edges;
 use Fhaculty\Graph\Set\Vertices;
 
+/**
+ * Abstract base class for shortest path algorithms
+ *
+ * This abstract base class provides the base interface for working with
+ * single-source shortest paths (SSSP).
+ *
+ * The shortest path problem is the problem of finding a path between two
+ * vertices such that the sum of the weights of its constituent edges is
+ * minimized. The weight of the shortest path is referred to as distance.
+ *
+ *    A--[10]-------------B---E<--F
+ *     \                 /
+ *      \--[4]--C--[2]--D
+ *
+ * In the above pictured graph, the distance (weight of the shortest path)
+ * between A and C is 4, and the shortest path between A and B is "A->C->D->B"
+ * with a distance (total weight) of 6.
+ *
+ * In graph theory, it is usually assumed that a path to an unreachable vertex
+ * has infinite distance. In the above pictured graph, there's no way path
+ * from A to F, i.e. vertex F is unreachable from vertex A because of the
+ * directed edge "E <- F" pointing in the opposite direction. This library
+ * considers this an Exception instead. So if you're asking for the distance
+ * between A and F, you'll receive an OutOfBoundsException instead.
+ *
+ * In graph theory, it is usually assumed that each vertex has a (pseudo-)path
+ * to itself with a distance of 0. In order to produce reliable, consistent
+ * results, this library considers this (pseudo-)path to be non-existant, i.e.
+ * there's NO "magic" path between A and A. So if you're asking for the distance
+ * between A and A, you'll receive an OutOfBoundsException instead. This allows
+ * us to check hether there's a real path between A and A (cycle via other
+ * vertices) as well as working with loop edges.
+ *
+ * @link http://en.wikipedia.org/wiki/Shortest_path_problem
+ * @link http://en.wikipedia.org/wiki/Tree_%28data_structure%29
+ * @see ShortestPath\Dijkstra
+ * @see ShortestPath\MooreBellmanFord which also supports negative Edge weights
+ * @see ShortestPath\BreadthFirst with does not consider Edge weights, but only the number of hops
+ */
 abstract class Base extends BaseVertex
 {
     /**
@@ -18,7 +57,7 @@ abstract class Base extends BaseVertex
      *
      * @param  Vertex    $endVertex
      * @return Walk
-     * @throws Exception when there's no walk from start to end vertex
+     * @throws OutOfBoundsException if there's no path to the given end vertex
      * @uses self::getEdgesTo()
      * @uses Walk::factoryFromEdges()
      */
@@ -31,10 +70,10 @@ abstract class Base extends BaseVertex
      * get array of edges (path) from start vertex to given end vertex
      *
      * @param  Vertex    $endVertex
-     * @throws Exception
+     * @throws OutOfBoundsException if there's no path to the given end vertex
      * @return Edges
-     * @uses AlgorithmSp::getEdges()
-     * @uses AlgorithmSp::getEdgesToInternal()
+     * @uses self::getEdges()
+     * @uses self::getEdgesToInternal()
      */
     public function getEdgesTo(Vertex $endVertex)
     {
@@ -46,15 +85,15 @@ abstract class Base extends BaseVertex
      *
      * @param  Vertex    $endVertex
      * @param  array     $edges     array of all input edges to operate on
-     * @throws Exception
+     * @throws OutOfBoundsException if there's no path to the given vertex
      * @return Edges
-     * @uses AlgorithmSp::getEdges() if no edges were given
+     * @uses self::getEdges() if no edges were given
      */
     protected function getEdgesToInternal(Vertex $endVertex, array $edges)
     {
         $currentVertex = $endVertex;
         $path = array();
-        while ($currentVertex !== $this->vertex) {
+        do {
             $pre = NULL;
             // check all edges to search for edge that points TO current vertex
             foreach ($edges as $edge) {
@@ -68,9 +107,9 @@ abstract class Base extends BaseVertex
                 } // ignore: this edge does not point TO current vertex
             }
             if ($pre === NULL) {
-                throw new UnderflowException('No edge leading to vertex');
+                throw new OutOfBoundsException('No edge leading to vertex');
             }
-        }
+        } while ($currentVertex !== $this->vertex);
 
         return new Edges(array_reverse($path));
     }
@@ -96,7 +135,7 @@ abstract class Base extends BaseVertex
      * get set of all Vertices the given start vertex has a path to
      *
      * @return Vertices
-     * @uses AlgorithmSp::getDistanceMap()
+     * @uses self::getDistanceMap()
      */
     public function getVertices()
     {
@@ -112,12 +151,30 @@ abstract class Base extends BaseVertex
     }
 
     /**
+     * checks whether there's a path from this start vertex to given end vertex
+     *
+     * @param  Vertex  $endVertex
+     * @return boolean
+     * @uses self::getEdgesTo()
+     */
+    public function hasVertex(Vertex $vertex)
+    {
+        try {
+            $this->getEdgesTo($vertex);
+        }
+        catch (OutOfBoundsException $e) {
+            return false;
+        }
+        return true;
+    }
+
+    /**
      * get map of vertex IDs to distance
      *
      * @return float[]
-     * @uses AlgorithmSp::getEdges()
-     * @uses AlgorithmSp::getEdgesToInternal()
-     * @uses AlgorithmSp::sumEdges()
+     * @uses self::getEdges()
+     * @uses self::getEdgesToInternal()
+     * @uses self::sumEdges()
      */
     public function getDistanceMap()
     {
@@ -126,7 +183,7 @@ abstract class Base extends BaseVertex
         foreach ($this->vertex->getGraph()->getVertices()->getMap() as $vid => $vertex) {
             try {
                 $ret[$vid] = $this->sumEdges($this->getEdgesToInternal($vertex, $edges));
-            } catch (UnderflowException $ignore) {
+            } catch (OutOfBoundsException $ignore) {
             } // ignore vertices that can not be reached
         }
 
@@ -138,9 +195,9 @@ abstract class Base extends BaseVertex
      *
      * @param  Vertex    $endVertex
      * @return float
-     * @throws Exception if given vertex is invalid or there's no path to given end vertex
-     * @uses AlgorithmSp::getEdgesTo()
-     * @uses AlgorithmSp::sumEdges()
+     * @throws OutOfBoundsException if there's no path to the given end vertex
+     * @uses self::getEdgesTo()
+     * @uses self::sumEdges()
      */
     public function getDistance(Vertex $endVertex)
     {
@@ -150,8 +207,43 @@ abstract class Base extends BaseVertex
     /**
      * create new resulting graph with only edges on shortest path
      *
+     * The resulting Graph will always represent a tree with the start vertex
+     * being the root vertex.
+     *
+     * For example considering the following input Graph with equal weights on
+     * each edge:
+     *
+     *     A----->F
+     *    / \     ^
+     *   /   \   /
+     *  /     \ /
+     *  |      E
+     *  |       \
+     *  |        \
+     *  B--->C<---D
+     *
+     * The resulting shortest path tree Graph will look like this:
+     *
+     *     A----->F
+     *    / \
+     *   /   \
+     *  /     \
+     *  |      E
+     *  |       \
+     *  |        \
+     *  B--->C    D
+     *
+     * Or by just arranging the Vertices slightly different:
+     *
+     *          A
+     *         /|\
+     *        / | \
+     *       B  E  \->F
+     *      /   |
+     *  C<-/    D
+     *
      * @return Graph
-     * @uses AlgorithmSp::getEdges()
+     * @uses self::getEdges()
      * @uses Graph::createGraphCloneEdges()
      */
     public function createGraph()
@@ -171,8 +263,6 @@ abstract class Base extends BaseVertex
     protected function getEdgesCheapestPredecesor(array $predecessor)
     {
         $vertices = $this->vertex->getGraph()->getVertices()->getMap();
-        // start vertex doesn't have a predecessor
-        unset($vertices[$this->vertex->getId()]);
 
         $edges = array();
         foreach ($vertices as $vid => $vertex) {