+++ lib/Fhaculty/Graph/Algorithm/ShortestPath/Dijkstra.php
@@ -6,18 +6,31 @@ use Fhaculty\Graph\Set\Edges;
 use Fhaculty\Graph\Exception\UnexpectedValueException;
 use \SplPriorityQueue;
 
+/**
+ * Commonly used Dijkstra's shortest path algorithm
+ *
+ * This is asymptotically the fastest known single-source shortest-path
+ * algorithm for arbitrary graphs with non-negative weights. If your Graph
+ * contains an Edge with negative weight, if will throw an
+ * UnexpectedValueException. Consider using (the slower) MooreBellmanFord
+ * algorithm instead.
+ *
+ * @link http://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
+ * @see MooreBellmanFord
+ */
 class Dijkstra extends Base
 {
     /**
      * get all edges on shortest path for this vertex
      *
      * @return Edges
+     * @throws UnexpectedValueException when encountering an Edge with negative weight
      */
     public function getEdges()
     {
         $totalCostOfCheapestPathTo  = Array();
         // start node distance
-        $totalCostOfCheapestPathTo[$this->vertex->getId()] = 0;
+        $totalCostOfCheapestPathTo[$this->vertex->getId()] = INF;
 
         // just to get the cheapest vertex in the correct order
         $cheapestVertex = new SplPriorityQueue();
@@ -29,6 +42,8 @@ class Dijkstra extends Base
 
         // mark vertices when their cheapest path has been found
         $usedVertices  = Array();
+
+        $isFirst = true;
 
         // Repeat until all vertices have been marked
         $totalCountOfVertices = $this->vertex->getGraph()->getNumberOfVertices();
@@ -53,8 +68,12 @@ class Dijkstra extends Base
                 break;
             }
 
-            // mark this vertex
-            $usedVertices[$currentVertexId] = true;
+            if ($isFirst) {
+                $isFirst = false;
+            } else {
+                // mark this vertex
+                $usedVertices[$currentVertexId] = true;
+            }
 
             // check for all edges of current vertex if there is a cheaper path (or IN OTHER WORDS: Add reachable nodes from currently added node and refresh the current possible distances)
             foreach ($currentVertex->getEdgesOut() as $edge) {
@@ -70,6 +89,9 @@ class Dijkstra extends Base
                 if (!isset($usedVertices[$targetVertexId])) {
                     // calculate new cost to vertex
                     $newCostsToTargetVertex = $totalCostOfCheapestPathTo[$currentVertexId] + $weight;
+                    if (is_infinite($newCostsToTargetVertex)) {
+                        $newCostsToTargetVertex = $weight;
+                    }
 
                     if ((!isset($predecesVertexOfCheapestPathTo[$targetVertexId]))
                            // is the new path cheaper?
@@ -87,6 +109,10 @@ class Dijkstra extends Base
                     }
                 }
             }
+        }
+
+        if ($totalCostOfCheapestPathTo[$this->vertex->getId()] === INF) {
+            unset($predecesVertexOfCheapestPathTo[$this->vertex->getId()]);
         }
 
         // algorithm is done, return resulting edges