+++ CHANGELOG.md
@@ -4,6 +4,35 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## 0.7.0 (2013-xx-xx)
+
+* BC break: Merge `Cycle` into `Walk` ([#61](https://github.com/clue/graph/issues/61)).
+As such, its static factory methods had to be renamed. Update your references if applicable:
+
+| Old name | New name |
+|---|---|
+| `Cycle::factoryFromPredecessorMap()` | `Walk::factoryCycleFromPredecessorMap()` |
+| `Cycle::factoryFromVertices()` | `Walk::factoryCycleFromVertices()` |
+| `Cycle::factoryFromEdges()` | `Walk::factoryCycleFromEdges()` |
+
+* BC break: Each of the above methods (`Walk::factoryCycleFromPredecessorMap()`,
+`Walk::factoryCycleFromVertices()`, `Walk::factoryCycleFromEdges()`) now
+actually makes sure the returned `Walk` instance is actually a valid Cycle,
+i.e. the start `Vertex` is the same as the end `Vertex` ([#61](https://github.com/clue/graph/issues/61))
+* BC break: Each `Algorithm\ShortestPath` algorithm now consistenly does not
+return a zero weight for the root Vertex and now supports loop edges on the root
+Vertex ([#62](https://github.com/clue/graph/issues/62))
+* BC break: Each `Algorithm\ShortestPath` algorithm now consistently throws an
+`OutOfBoundsException` for unreachable vertices
+([#62](https://github.com/clue/graph/issues/62))
+* Feature: Add `Algorithm\ShortestPath::hasVertex(Vertex $vertex)` to check whether
+a path to the given Vertex exists ([#62](https://github.com/clue/graph/issues/62)).
+* Fix: Checking `Walk::isValid()` ([#61](https://github.com/clue/graph/issues/61))
+* Fix: Missing import prevented
+`Algorithm\ShortestPath\MooreBellmanFord::getCycleNegative()` from actually
+throwing the right `UnderflowException` if no cycle was actually found
+([#62](https://github.com/clue/graph/issues/62))
+
 ## 0.6.0 (2013-07-11)
 
 * BC break: Move algorithm definitions in base classes to separate algorithm classes ([#27](https://github.com/clue/graph/issues/27)).